+++ lib/investigationCollection/index.ts
@@ -3,3 +3,4 @@ export * from "./dnsProvider";
 export * from "./orchestrator";
 export * from "./capabilityRegistry";
 export * from "./publicWebProvider";
+export * from "./secEdgarProvider";